Moved structs to `game_window`
Closes https://github.com/PistonDevelopers/piston/issues/566
